updated files with too many functions, some norm updates
