+++ src/draw_3d.c
@@ -6,56 +6,12 @@
 /*   By: jdagoy <jdagoy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/02 00:24:06 by jdagoy            #+#    #+#             */
-/*   Updated: 2024/01/19 09:19:34 by jdagoy           ###   ########.fr       */
+/*   Updated: 2024/01/19 10:09:30 by jdagoy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 #include "enums.h"
-
-static float	fix_fisheye(t_game *game)
-{
-	float	angle;
-
-	angle = game->player.angle - game->ray.angle;
-	if (angle < 0)
-		return (angle += 2 * M_PI);
-	if (angle > 2 * M_PI)
-		return (angle -= 2 * M_PI);
-	return (angle);
-}
-
-static int	get_direction(t_game *game)
-{
-	if (game->ray.hit == 1)
-	{
-		if (game->ray.angle > M_PI)
-			return (SOUTH);
-		else
-			return (NORTH);
-	}
-	else if (game->ray.angle > M_PI_2 && game->ray.angle < M_PI_3)
-		return (WEST);
-	else
-		return (EAST);
-}
-
-static float	check_rayhit(t_game *game, float x)
-{
-	if (game->ray.hit == 1)
-	{
-		x = (int)(game->ray.x / 3) % TILE_SIZE;
-		if (get_direction(game) == SOUTH)
-			return (TILE_SIZE - x - 1);
-	}
-	else
-	{
-		x = (int)(game->ray.y / 2) % TILE_SIZE;
-		if (get_direction(game) == WEST)
-			return (TILE_SIZE - x - 1);
-	}
-	return (x);
-}
 
 static int	get_texturecolor(t_game *game, t_walltexture *texture)
 {
