+++ src/process_map.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   process_map.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jdagoy <jdagoy@student.s19.be>             +#+  +:+       +#+        */
+/*   By: jdagoy <jdagoy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/18 23:55:06 by jdagoy            #+#    #+#             */
-/*   Updated: 2024/01/18 23:55:23 by jdagoy           ###   ########.fr       */
+/*   Updated: 2024/01/19 10:06:25 by jdagoy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-static	int exit_copymap(char **temp, int i)
+static int	exit_copymap(char **temp, int i)
 {
 	while (--i >= 0)
 		free(temp[i]);
@@ -22,10 +22,10 @@ static	int exit_copymap(char **temp, int i)
 
 int	process_map(t_game *game, char **map)
 {
-	char 	**temp;
+	char	**temp;
 	int		i;
 	int		j;
-	
+
 	temp = (char **)ft_calloc(sizeof(char *), game->map_height);
 	if (temp == NULL)
 		return (EXIT_FAILURE);
