+++ src/utils.c
@@ -10,22 +10,6 @@ int	len_split(char **split)
 	return (i);
 }
 
-void	ft_freesplit(char **tab)
-{
-	int	i;
-	int	len;
-
-	if (!tab)
-		return ;
-	i = -1;
-	len = len_split(tab);
-	while (++i < len)
-		if (tab[i])
-			free(tab[i]);
-	free(tab);
-}
-
-
 int	rgb_to_hex(int red, int green, int blue)
 {
 	int	res;
@@ -54,9 +38,9 @@ char	*skip_empty_line(int fd, char *line)
 	return (line);
 }
 
-void replace_threes(char **map)
+void	replace_threes(char **map)
 {
-	int i, j;
+	int	i, j;
 	i = 0;
 
 	while (map[i])
@@ -73,8 +57,7 @@ void replace_threes(char **map)
 	}
 }
 
-
-char *next_line(int fd, char *line)
+char	*next_line(int fd, char *line)
 {
 	free(line);
 	return (get_next_line(fd));
